requirements.txt
mlx-whisper
pyannote.audio
demucs
librosa
pyacoustid
opencv-python
